Refactored message parsing code into a separate MessageParser class
